moved classes to model
