+++ src/main/java/com/example/event_manager/service/EventServiceImpl.java
@@ -10,12 +10,12 @@ import com.example.event_manager.form.TaskStatusForm;
 import com.example.event_manager.mapper.EventMapper;
 import com.example.event_manager.mapper.EventWithBillingMapper;
 import com.example.event_manager.mapper.TaskStatusMapper;
+import com.example.event_manager.model.BillingRaportSchema;
+import com.example.event_manager.model.BillingsSummary;
 import com.example.event_manager.model.Event;
 import com.example.event_manager.model.TaskStatus;
 import com.example.event_manager.repo.EventRepo;
 import com.example.event_manager.utils.BillingRaport;
-import com.example.event_manager.utils.BillingRaportSchema;
-import com.example.event_manager.utils.BillingsSummary;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.HashMap;
