+++ src/main/resources/static/newStyle.xsl
